api/requirements.txt
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
zope.interface==5.1.0
